requirements.txt
streamlit
python-dotenv
openai
Pillow
requests
newspaper3k
textblob
